Address PR feedback
Re-work the return types - there is now a base interface with attributes
common to both CPU and GPU API usage, and each specific mode of
accessing the data has its own interface that inherits from the base one
& introduces usage-specific attributes / methods.

This API shape should be easily extensible to add WebGPU when it
arrives. It also allows us to relax the requirement that the XRFrame's
and XRWebGLBinding's getDepthInformation() methods throw when used in
different mode.