Fix explainer code and add validation in GPU path
Explainer relied on the sample code that worked due to a bug in Chrome's
implementation of depth sensing API. The type of `data` set on
`XRCPUDepthInformation` does not have the `buffer`, `byteOffset` and
`byteLength` attributes.

Additionally, add validation for the algorithm for
XRWebGLBinding.getDepthInformation() to ensure that the session of the
provided `XRView` matches the session of the binding itself.